layer working and minur bug fixed

+++ Code/keyboard.h
@@ -15,22 +15,22 @@ typedef struct
 {
     uint64_t t_time;
     uint8_t stat;
-    uint32_t keyvalue;
-    uint8_t charValue;
+    uint16_t keyvalue;
+    uint16_t charValue;
 } KeyObject;
 void keyboard_hander();
 void init_keyboard(uint8_t i_usbEN);
 
-uint8_t correspondKey(uint8_t key);
+uint16_t correspondKey(uint8_t key);
 void readkey();
 void settimestamp(uint8_t key,uint8_t key_status);
 
-uint8_t change_layer(volatile uint8_t **nextLayer,uint8_t key_char);
+uint8_t change_layer(volatile uint16_t **nextLayer,uint16_t key_char);
 void key_translate();
 
 void ringbuff_tail_plus_one(volatile ringBuff * i_ring); 
 
 void ringbuff_plus_one_head(volatile ringBuff * i_ring); 
 
-uint8_t get_char_from_layer(volatile uint8_t *ptr,uint8_t i_key,uint8_t i_method);
+uint8_t get_char_from_layer(volatile uint16_t *ptr,uint16_t i_key,uint8_t i_method);
 #endif
